fix updates

+++ es-app/src/SaveState.cpp
@@ -33,6 +33,23 @@ std::string SaveState::makeStateFilename(int slot, bool fullPath, bool useImageG
 	return ret;
 }
 
+std::string SaveState::makeAutoSaveFilename(bool fullPath, bool useImageGenerator) const
+{
+	std::string ret = useImageGenerator ? autosaveImageGenerator : autosaveFileGenerator;
+	if (ret.empty())
+		ret = useImageGenerator ? imageGenerator : fileGenerator;
+
+	ret = Utils::String::replace(ret, "{{slot}}", ".auto");
+	ret = Utils::String::replace(ret, "{{slot0}}", ".auto");
+	ret = Utils::String::replace(ret, "{{slot00}}", ".auto");
+	ret = Utils::String::replace(ret, "{{slot2d}}", ".auto");
+
+	if (fullPath)
+		return Utils::FileSystem::combine(Utils::FileSystem::getParent(fileName), ret);
+
+	return ret;
+}
+
 std::string SaveState::getScreenShot() const
 {
   return screenshot;
@@ -69,6 +86,11 @@ std::string SaveState::setupSaveState(FileData* game, const std::string& command
 	if (game == nullptr)
 		return command;
 
+	mAutoFileBackup.clear();
+	mAutoImageBackup.clear();
+	mNewSlotFile.clear();
+	mNewSlotCheckSum.clear();
+
 	std::string cmd = command;
 	auto system = game->getSourceFileData()->getSystem();
 
@@ -101,9 +123,43 @@ std::string SaveState::setupSaveState(FileData* game, const std::string& command
 		return cmd;
 	}
 
-	bool incrementalSaveStates = SystemConf::getIncrementalSaveStates() && /*hasAutosave && */supportsIncremental;
+	if (forceFixedSlotAutosave)
+	{
+		if (racommands)
+			return cmd + " -autosave 1 -state_slot " + std::to_string(slot);
 
-	std::string path = Utils::FileSystem::getParent(fileName);
+		if (!fileName.empty())
+			return cmd + " -state_slot " + std::to_string(slot) + " -state_file \"" + fileName + "\"";
+
+		return cmd + " -state_slot " + std::to_string(slot);
+	}
+
+	bool incrementalSaveStates = SystemConf::getIncrementalSaveStates() && /*hasAutosave && */supportsIncremental;
+	std::string autoFilename = makeAutoSaveFilename();
+	std::string autoImage = makeAutoSaveFilename(true, true);
+	bool hydrateAutosave = racommands && !fileName.empty() && !autoFilename.empty() && fileName != autoFilename;
+
+	if (hydrateAutosave)
+	{
+		if (Utils::FileSystem::exists(autoFilename))
+		{
+			Utils::FileSystem::removeFile(autoFilename + ".bak");
+			Utils::FileSystem::renameFile(autoFilename, autoFilename + ".bak");
+		}
+
+		if (Utils::FileSystem::exists(autoImage))
+		{
+			Utils::FileSystem::removeFile(autoImage + ".bak");
+			Utils::FileSystem::renameFile(autoImage, autoImage + ".bak");
+		}
+
+		mAutoImageBackup = autoImage;
+		mAutoFileBackup = autoFilename;
+
+		Utils::FileSystem::copyFile(fileName, autoFilename);
+		if (!getScreenShot().empty())
+			Utils::FileSystem::copyFile(getScreenShot(), autoImage);
+	}
 
 	if (slot == -1) // Run current AutoSave
 	{
@@ -147,37 +203,13 @@ std::string SaveState::setupSaveState(FileData* game, const std::string& command
 
 		if (racommands) 
 		{
-			// Copy to state.auto file
-			auto autoFilename = makeStateFilename(-1);
-			if (Utils::FileSystem::exists(autoFilename))
+			if (!fileName.empty() && incrementalSaveStates && nextSlot >= 0 && slot + 1 != nextSlot)
 			{
-				Utils::FileSystem::removeFile(autoFilename + ".bak");
-				Utils::FileSystem::renameFile(autoFilename, autoFilename + ".bak");
-			}
-
-			// Copy to state.auto.png file
-			auto autoImage = autoFilename + ".png";
-			if (Utils::FileSystem::exists(autoImage))
-			{
-				Utils::FileSystem::removeFile(autoImage + ".bak");
-				Utils::FileSystem::renameFile(autoImage, autoImage + ".bak");
-			}
-
-			mAutoImageBackup = autoImage;
-			mAutoFileBackup = autoFilename;
-
-			if (!fileName.empty())
-			{
-				Utils::FileSystem::copyFile(fileName, autoFilename);
-
-				if (incrementalSaveStates && nextSlot >= 0 && slot + 1 != nextSlot)
-				{
-					// Copy file to new slot, if the users want to reload the saved game in the slot directly from retroach
-					mNewSlotFile = makeStateFilename(nextSlot);
-					Utils::FileSystem::removeFile(mNewSlotFile);
-					if (Utils::FileSystem::copyFile(fileName, mNewSlotFile))
-						mNewSlotCheckSum = ApiSystem::getInstance()->getMD5(fileName, false);
-				}
+				// Copy file to new slot, if the users want to reload the saved game in the slot directly from retroarch.
+				mNewSlotFile = makeStateFilename(nextSlot);
+				Utils::FileSystem::removeFile(mNewSlotFile);
+				if (Utils::FileSystem::copyFile(fileName, mNewSlotFile))
+					mNewSlotCheckSum = ApiSystem::getInstance()->getMD5(fileName, false);
 			}
 		}
 	}
@@ -199,17 +231,21 @@ void SaveState::onGameEnded(FileData* game)
 			if (fileChecksum == mNewSlotCheckSum)
 				Utils::FileSystem::removeFile(mNewSlotFile);
 		}
+		mNewSlotFile.clear();
+		mNewSlotCheckSum.clear();
 
 		if (!mAutoFileBackup.empty())
 		{
 			Utils::FileSystem::removeFile(mAutoFileBackup);
 			Utils::FileSystem::renameFile(mAutoFileBackup + ".bak", mAutoFileBackup);
+			mAutoFileBackup.clear();
 		}
 
 		if (!mAutoImageBackup.empty())
 		{
 			Utils::FileSystem::removeFile(mAutoImageBackup);
 			Utils::FileSystem::renameFile(mAutoImageBackup + ".bak", mAutoImageBackup);
+			mAutoImageBackup.clear();
 		}
 	}
